+++ src/components/beta/gux-error-message-beta/gux-error-message-beta.tsx
@@ -2,7 +2,7 @@ import { Component, Element, h } from '@stencil/core';
 import { trackComponent } from '../../../usage-tracking';
 
 /**
- * @slot error - Required slot for error
+ * @slot error(default) - Required slot for error
  */
 @Component({
   styleUrl: 'gux-error-message-beta.less',
@@ -22,7 +22,8 @@ export class GuxErrorMessageBeta {
       <div class="gux-container">
         <gux-icon icon-name="alert-warning-octogon" decorative></gux-icon>
         <div class="gux-message">
-          <slot slot="error" />
+          <slot />
+          <slot name="error" />
         </div>
       </div>
     );