+++ src/components/stable/gux-datepicker/gux-datepicker.tsx
@@ -13,13 +13,15 @@ import {
 
 import { trackComponent } from '../../../usage-tracking';
 import { CalendarModes } from '../../../common-enums';
+import { buildI18nForComponent, GetI18nValue } from '../../../i18n';
+
 import {
   asIsoDateRange,
   asIsoDate,
   fromIsoDateRange,
   fromIsoDate
 } from '../../../utils/date/iso-dates';
-import { buildI18nForComponent, GetI18nValue } from '../../../i18n';
+import { OnClickOutside } from '../../../utils/decorator/on-click-outside';
 
 import translationResources from './i18n/en.json';
 import {
@@ -48,6 +50,7 @@ import {
 export class GuxDatepicker {
   yearFormat: string = 'yyyy';
   intervalOrder: string[];
+  datepickerElement: HTMLElement;
   inputElement: HTMLInputElement;
   toInputElement: HTMLInputElement;
   calendarElement: HTMLGuxCalendarElement;
@@ -252,7 +255,6 @@ export class GuxDatepicker {
         this.format,
         getIntervalLetter(this.format, 0)
       );
-      this.active = false;
     }
   }
 
@@ -260,6 +262,20 @@ export class GuxDatepicker {
   onMouseDown(event: MouseEvent) {
     if (this.isFocusedFieldEvent(event)) {
       this.isSelectingRange = true;
+    }
+
+    const target = event.target as HTMLElement;
+    const inDatepicker = this.datepickerElement.contains(target);
+
+    const notToggleButton = Array.from(
+      this.root.querySelectorAll('.gux-calendar-toggle-button')
+    ).every(
+      (toggleButtonElement: HTMLButtonElement) =>
+        !toggleButtonElement.contains(target)
+    );
+
+    if (notToggleButton) {
+      this.active = inDatepicker;
     }
   }
 
@@ -275,6 +291,11 @@ export class GuxDatepicker {
       );
       this.setRange();
     }
+  }
+
+  @OnClickOutside({ triggerEvents: 'mousedown' })
+  onClickOutside() {
+    this.active = false;
   }
 
   isFocusedFieldEvent(event: Event): boolean {
@@ -587,6 +608,7 @@ export class GuxDatepicker {
   renderCalendarToggleButton(): JSX.Element {
     return (
       <button
+        class="gux-calendar-toggle-button"
         aria-hidden="true"
         type="button"
         onClick={() => this.toggleCalendar()}
@@ -664,6 +686,7 @@ export class GuxDatepicker {
           'gux-datepicker': true,
           'gux-active': this.active
         }}
+        ref={(el: HTMLElement) => (this.datepickerElement = el)}
       >
         {this.renderStartDateField()}
         {this.renderEndDateField()}