+++ src/components/stable/gux-dropdown/gux-dropdown.tsx
@@ -3,6 +3,7 @@ import {
   Element,
   Event,
   EventEmitter,
+  forceUpdate,
   h,
   Listen,
   Method,
@@ -10,9 +11,9 @@ import {
   State
 } from '@stencil/core';
 
-import { KeyCode } from '../../../common-enums';
 import { whenEventIsFrom } from '../../../utils/dom/when-event-is-from';
 import { trackComponent } from '../../../usage-tracking';
+import { OnMutation } from '../../../utils/decorator/on-mutation';
 
 @Component({
   styleUrl: 'gux-dropdown.less',
@@ -22,7 +23,7 @@ export class GuxDropdown {
   @Element()
   root: HTMLElement;
 
-  textFieldElement: HTMLGuxTextFieldLegacyElement;
+  textFieldElement: HTMLInputElement;
 
   /**
    * Sets the select mode (default, page or palette).
@@ -88,8 +89,8 @@ export class GuxDropdown {
   }
 
   @Method()
-  async setLabeledBy(labeledBy: string): Promise<void> {
-    this.textFieldElement.setLabelledBy(labeledBy);
+  async setLabeledBy(id: string): Promise<void> {
+    this.srLabeledBy = id;
   }
 
   @Method()
@@ -117,17 +118,23 @@ export class GuxDropdown {
     this.value = '';
   }
 
+  @OnMutation({ childList: true, subtree: true })
+  onMutation(): void {
+    forceUpdate(this.root);
+    this.setSelected();
+  }
+
   // TODO: Fix the keyboard navigation I broke
   onKeyDown(event: KeyboardEvent) {
     const selectionOptions = this.getSelectionOptions();
     const focusIndex = this.getFocusIndex(selectionOptions);
-    switch (event.keyCode) {
-      case KeyCode.Up:
+    switch (event.key) {
+      case 'ArrowUp':
         if (focusIndex > 0) {
           selectionOptions[focusIndex - 1].focus();
         }
         break;
-      case KeyCode.Down:
+      case 'ArrowDown':
         if (this.inputIsFocused) {
           this.opened = true;
         }
@@ -135,20 +142,20 @@ export class GuxDropdown {
           selectionOptions[focusIndex + 1].focus();
         }
         break;
-      case KeyCode.Home:
+      case 'Home':
         if (!selectionOptions.length) {
           return;
         }
         selectionOptions[0].focus();
         break;
-      case KeyCode.End:
+      case 'End':
         if (!selectionOptions.length) {
           return;
         }
         selectionOptions[selectionOptions.length - 1].focus();
         break;
-      case KeyCode.Enter:
-      case KeyCode.Space:
+      case 'Enter':
+      case 'Space':
         break;
       default:
         this.valueEdited = true;
@@ -188,19 +195,28 @@ export class GuxDropdown {
     });
   }
 
-  private _optionClickedHandler(e: MouseEvent) {
+  private optionSelectedHandler(e: Event) {
     whenEventIsFrom('gux-option', e, elem => {
       const option = elem as HTMLGuxOptionElement;
-      this.setValue(option.text, option.value || option.text);
+      const selectionOptions = this.getSelectionOptions();
+
+      selectionOptions.forEach(selectionOption => {
+        if (selectionOption === option) {
+          selectionOption.selected = true;
+          this.setValue(
+            selectionOption.text,
+            selectionOption.value || selectionOption.text
+          );
+        } else {
+          selectionOption.selected = false;
+        }
+      });
     });
   }
 
-  private _optionKeyDownHandler(e: KeyboardEvent) {
+  private optionKeyDownHandler(e: KeyboardEvent) {
     if (e.key === ' ' || e.key === 'Enter') {
-      whenEventIsFrom('gux-option', e, elem => {
-        const option = elem as HTMLGuxOptionElement;
-        this.setValue(option.text, option.value || option.text);
-      });
+      this.optionSelectedHandler(e);
     }
   }
 
@@ -209,18 +225,9 @@ export class GuxDropdown {
     this.forcedGhostValue = '';
   }
 
-  _inputHandler(event: CustomEvent) {
-    this.value = event.detail;
+  _inputHandler(inputEvent: Event) {
+    this.value = (inputEvent.target as HTMLInputElement).value;
     this.opened = true;
-  }
-
-  private _showDropdownIcon() {
-    const selectionOptions = this.getSelectionOptions();
-    const match = selectionOptions.filter(item => {
-      return item.text === this.value;
-    });
-    const filterableBehavior = !this.value || !!match.length;
-    return this.filterable ? filterableBehavior : true;
   }
 
   getFilteredItems() {
@@ -235,16 +242,21 @@ export class GuxDropdown {
     }
   }
 
-  getGhost() {
+  getSuggestionText(filter: string = '') {
     this.searchHighlightAndFilter(this.value);
+    const filterLength = filter.length;
     const firstFilteredItem = this.getFilteredItems().length
       ? this.getFilteredItems()[0].text
       : '';
-    const valueGhost =
-      this.value + firstFilteredItem.substring(this.value.length);
-    const ghost = this.forcedGhostValue ? this.forcedGhostValue : valueGhost;
-    const placeholder = !this.value ? this.placeholder : '';
-    return this.opened && this.filterable ? ghost : placeholder;
+    if (filterLength > 0) {
+      const option = firstFilteredItem;
+      if (option) {
+        return this.opened && this.filterable
+          ? option.substring(filterLength)
+          : '';
+      }
+    }
+    return '';
   }
 
   componentWillLoad(): void {
@@ -257,7 +269,7 @@ export class GuxDropdown {
     this.setSelected();
 
     if (!this.filterable) {
-      this.textFieldElement.readonly = true;
+      this.textFieldElement.readOnly = true;
     }
   }
 
@@ -290,46 +302,52 @@ export class GuxDropdown {
         onKeyDown={e => this.onKeyDown(e)}
       >
         <div class="gux-select-field">
-          <span class="gux-ghost" aria-hidden="true">
-            {this.getGhost()}
+          <span
+            class={`gux-filter-suggestion ${this.opened ? 'gux-active' : ''}`}
+            aria-hidden="true"
+          >
+            <span class="gux-filter-text">{this.value}</span>
+            <span class="gux-filter-typeahead">
+              {this.getSuggestionText(this.value)}
+            </span>
           </span>
-          <gux-text-field-legacy
-            title={this.value}
-            ref={el =>
-              (this.textFieldElement = el as HTMLGuxTextFieldLegacyElement)
-            }
-            onMouseDown={() => {
-              this._clickHandler();
-            }}
-            onFocus={() => {
-              this._focusHandler();
-            }}
-            onBlur={() => {
-              this._blurHandler();
-            }}
-            onInput={e => {
-              this._inputHandler(e);
-            }}
-            value={this.value}
-            disabled={this.disabled}
-            class={this._showDropdownIcon() ? 'gux-unclearable' : ''}
-          />
-          {this._showDropdownIcon() && (
-            <button
-              class="gux-dropdown-indicator"
-              aria-hidden="true"
-              tabindex="-1"
-              type="button"
-            >
-              <gux-icon decorative icon-name="chevron-small-down"></gux-icon>
-            </button>
-          )}
+
+          <gux-input-text-like>
+            <input
+              placeholder={this.placeholder}
+              slot="input"
+              value={this.value}
+              aria-labelledby={this.srLabeledBy}
+              ref={ref => (this.textFieldElement = ref)}
+              onMouseDown={() => {
+                this._clickHandler();
+              }}
+              onFocus={() => {
+                this._focusHandler();
+              }}
+              onBlur={() => {
+                this._blurHandler();
+              }}
+              onInput={e => {
+                this._inputHandler(e);
+              }}
+            />
+          </gux-input-text-like>
+
+          <button
+            class="gux-dropdown-indicator"
+            aria-hidden="true"
+            tabindex="-1"
+            type="button"
+          >
+            <gux-icon decorative icon-name="chevron-small-down"></gux-icon>
+          </button>
         </div>
         <div
           class={`gux-options ${this.opened ? 'gux-opened' : ''}`}
-          onClick={this._optionClickedHandler.bind(this)}
+          onClick={this.optionSelectedHandler.bind(this)}
           onFocusin={this._optionFocusedHandler.bind(this)}
-          onKeyDown={this._optionKeyDownHandler.bind(this)}
+          onKeyDown={this.optionKeyDownHandler.bind(this)}
         >
           <slot />
         </div>